requirements.txt
# ── Core Framework ──────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
streamlit==1.35.0
python-multipart==0.0.9

# ── AI / LLM ─────────────────────────────────────────────────────────────────
anthropic==0.28.0

# ── Job Scraping ─────────────────────────────────────────────────────────────
python-jobspy==1.1.65
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.21.0
webdriver-manager==4.0.1

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9      # swap to psycopg2 in prod
aiosqlite==0.20.0            # SQLite async (dev)

# ── Document Generation ───────────────────────────────────────────────────────
python-docx==1.1.2
reportlab==4.2.0
Pillow==10.3.0

# ── Resume Parsing ────────────────────────────────────────────────────────────
pdfminer.six==20221105
PyMuPDF==1.24.3              # fast PDF text extraction

# ── Data / Utils ─────────────────────────────────────────────────────────────
pandas==2.2.2
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
tenacity==8.3.0

# ── NLP / Scoring ─────────────────────────────────────────────────────────────
scikit-learn==1.5.0
nltk==3.8.1
